tune: Add for loop for project generation (initial prototype)

+++ test/torture.cpp
@@ -17,7 +17,7 @@
 #include "test_impl.hpp"
 
 int main() {
-  int result = 0;
+  size_t result = 0;
   for (unsigned i = 0; i < 10000000; ++i) {
     result += make_some_stacktrace1()[0].source_line();
   }
